Move: 타입을 따로 관리하기 위해 폴더 및 파일 생성
