+++ src/components/LoadingWrapper.tsx
@@ -2,13 +2,11 @@
 
 import React, { useEffect, useState } from 'react';
 
+import { LoadingWrapperProps } from '@/types/loadingWrapperPropsType';
+
 import InitialLoadingScreen from '@/components/InitialLoadingScreen';
 import Navbar from './Navbar';
 import Footer from './Footer';
-
-interface LoadingWrapperProps {
-  children: React.ReactNode;
-}
 
 const LoadingWrapper = ({ children }: LoadingWrapperProps) => {
   const [isLoading, setIsLoading] = useState(false);
