+++ src/types/loadingWrapperPropsType.ts
@@ -1,0 +1,3 @@
+export interface LoadingWrapperProps {
+  children: React.ReactNode;
+}
