+++ src/types/providerPropsType.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface ProviderProps {
+  children: ReactNode;
+}
